♻️Refactoring: EXIF extraction moved to the dedicated class
